refactor: move analytics interface definitions from service to dashboard models and add by country filter
